packages/api/pyproject.toml
[tool.poetry]
name = "eudi-connect"
version = "0.1.0"
description = "EUDI-Connect API - EU Digital Identity Wallet Integration Layer"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "eudi_connect"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.22"}
asyncpg = "^0.28.0"
pydantic = {extras = ["email"], version = "^2.4.2"}
pydantic-settings = "^2.0.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
httpx = "^0.27.0"
opentelemetry-api = "1.11.1"
opentelemetry-sdk = "1.11.1"
opentelemetry-instrumentation-fastapi = "0.30b1"
opentelemetry-exporter-otlp = "1.11.1"
didkit = "^0.3.3"
stripe = "^7.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
pytest-benchmark = "^4.0.0"
black = "^23.10.1"
isort = "^5.12.0"
mypy = "^1.6.1"
ruff = "^0.1.3"
locust = "^2.17.0"
streamlit = "^1.29.0"
plotly = "^5.18.0"
pandas = "^2.1.4"
slack-sdk = "^3.26.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]
